+++ linear_programming.h
@@ -5,8 +5,8 @@
 
 using namespace std;
 
-double one_d_linear(const vector<double> &a, const vector<double> &b, double c, int *flag, int *index);
+long double one_d_linear(const vector<long double> &a, const vector<long double> &b, long double c, int *flag, int *index);
 
-vector<point> find_bridge(vector<point> &p, double pivot, int *flag);
+vector<point> find_bridge(vector<point> &p, long double pivot, int *flag);
 
 #endif
